feat: refactor landing page
